codex-review: linux: cast voltage to signed before multiplying by current
CR-026 (3/10 reviewers, P2). The Class A sign-normalization step
introduced a CURRENT_NOW path that multiplies a signed int64_t
(batteryCurrent, can be negative) by a uint64_t (batteryVoltage).
C arithmetic conversions promote the signed operand to unsigned before
the multiplication, so a negative current wraps to a huge positive
value. The later abs() in the STATUS-based sign normalization then
sees the wrapped value and publishes a wildly bogus powerCurr.

Cast batteryVoltage to int64_t for the multiplication. Voltage in
microvolts is well within int64_t range; product (current * voltage,
worst case ~6.5e13) is far below int64_t max (~9.2e18), so no overflow
risk. The other charge*voltage multiplications on lines 1037 and 1040
involve uint64_t * uint64_t and are unaffected.

+++ linux/Platform.c
@@ -1066,7 +1066,7 @@ static void Platform_Battery_getSysData(BatteryInfo* info) {
             unitsContributingCharge++;
 
          if (!haveBatteryPower && haveBatteryCurrent && haveBatteryVoltage) {
-            batteryPower = (batteryCurrent * batteryVoltage) / 1000000;
+            batteryPower = (batteryCurrent * (int64_t)batteryVoltage) / 1000000;
             haveBatteryPower = true;
          }
